+++ packages/database-manager/test/unit/user-repository-test.ts
@@ -8,7 +8,7 @@ let userRepo: UserRepository;
 
 describe('user-test', () => {
   beforeEach(() => {
-    userRepo = new UserRepository('user_test');
+    userRepo = new UserRepository();
     testUser = new TestUser();
     testUser.login = 'login';
     testUser.password = '123';
@@ -25,43 +25,52 @@ describe('user-test', () => {
     const isAdded = await userRepo.addUser(testUser);
     const fromDatabaseUser = await userRepo.getUser('login');
     await userRepo.removeUser('login');
-    expect(isAdded).equal(true);
-    expect(fromDatabaseUser.login).equal(testUser.login);
-    expect(fromDatabaseUser.password).equal(testUser.password);
-    expect(fromDatabaseUser.isAdmin).equal(testUser.isAdmin);
+    expect(isAdded).equal(0);
+    expect(fromDatabaseUser).not.equal(400);
+    if (fromDatabaseUser !== 400) {
+      expect(fromDatabaseUser.login).equal(testUser.login);
+      expect(fromDatabaseUser.password).equal(testUser.password);
+      expect(fromDatabaseUser.isAdmin).equal(testUser.isAdmin);
+    }
   }).timeout(5000);
 
   it('user-get-error-test', async () => {
-    await expect(userRepo.getUser('a')).to.be.rejectedWith(Error);
+    expect(await userRepo.getUser('a')).equal(400);
   }).timeout(5000);
 
   it('user-remove-test', async () => {
     await userRepo.addUser(testUser);
     const isRemoved = await userRepo.removeUser('login');
-    expect(isRemoved).equal(true);
-    await expect(userRepo.getUser('a')).to.be.rejectedWith(Error);
+    expect(isRemoved).equal(0);
+    expect(await userRepo.getUser('a')).equal(400);
   }).timeout(5000);
 
   it('user-remove-false-test', async () => {
     const isRemoved = await userRepo.removeUser('a');
-    expect(isRemoved).equal(false);
+    expect(isRemoved).equal(400);
   }).timeout(5000);
 
   it('user-update-test', async () => {
     await userRepo.addUser(testUser);
     const beforeUpdateUser = await userRepo.getUser('login');
-    expect(beforeUpdateUser.isAdmin).equal(false);
-    testUser.isAdmin = true;
-    const updated = await userRepo.updateUser(testUser);
-    const updatedUser = await userRepo.getUser('login');
-    await userRepo.removeUser('login');
-    expect(updated).equal(true);
-    expect(updatedUser.isAdmin).equal(true);
+    expect(beforeUpdateUser).not.equal(400);
+    if (beforeUpdateUser !== 400) {
+      expect(beforeUpdateUser.isAdmin).equal(false);
+      testUser.isAdmin = true;
+      const updated = await userRepo.updateUser(testUser);
+      const updatedUser = await userRepo.getUser('login');
+      await userRepo.removeUser('login');
+      expect(updated).equal(0);
+      expect(updatedUser).not.equal(400);
+      if (updatedUser !== 400) {
+        expect(updatedUser.isAdmin).equal(true);
+      }
+    }
   }).timeout(5000);
 
   it('user-update-false-test', async () => {
     const updated = await userRepo.updateUser(testUser);
-    expect(updated).equal(false);
+    expect(updated).equal(400);
   }).timeout(5000);
 });
